+++ example/basic/src/services/dataService.ts
@@ -1,4 +1,4 @@
-import { Injectable } from "../../../dist"
+import { Injectable } from "../../../../dist"
 
 @Injectable()
 export default class dataService {