+++ lib/decorators/index.ts
@@ -1,5 +1,4 @@
 export * from "./Controller";
 export * from "./Injectable";
-export * from "./Middleware";
 export * from "./Resolver";
 export * from "./Route";