+++ lib/interface/index.ts
@@ -1,5 +1,3 @@
-import { Handler } from '../';
-
 /**
  * Represents a resolver Interface
  * @typedef {Object} IResolver
@@ -7,13 +5,4 @@ import { Handler } from '../';
  */
 export interface IResolver<T = unknown> {
     resolve: ((param?: string) => T) | ((param?: string) => Promise<T>) ;
-}
-
-/**
- * Represents a middleware Interface
- * @typedef {Object} IMiddleware
- * @property {Handler} execute - The execute function of the middleware
- */
-export interface IMiddleware {
-    execute: Handler
 }